Major rework - 1 on viewServer
