--- central-server/docker-compose.yml
@@ -1,46 +1,0 @@
-version: "3"
-services:
-  multicluster-central-database:
-    restart: "no"
-    image: postgres:10.8-alpine
-    container_name: multicluster-central-database
-    environment:
-      POSTGRES_USER: postgres
-      POSTGRES_DB: multicluster_db
-    ports:
-      - "4300:5432"
-    networks:
-      - multicluster
-    # volumes:
-    #   - db-data:/var/lib/postgresql/data
-  multicluster-central-api:
-    restart: "no"
-    build:
-      context: .
-      dockerfile: Dockerfile
-    container_name: multicluster-central-api
-    environment:
-      DATABASE_URI: postgresql://postgres:postgres@multicluster-central-database:5432/multicluster_db
-      FLASK_APP_SECRET: qY2i691SX2sEuZ7LUjY480RS48mw5qCeUiyV0i0vzmg
-      FLASK_ENV: development
-      DATABASE_USER: postgres
-      FLASK_APP: server.py
-      FLASK_RUN_PORT: 5000
-    ports:
-      - "5500:5000"
-    volumes:
-      - .:/app
-    links:
-      - multicluster-central-database
-    depends_on:
-      - multicluster-central-database
-    networks:
-      - multicluster
-
-# run on network called multicluster
-networks:
-  multicluster:
-    external: true
-
-# volumes:
-#   db-data:
