--- peer-app/docker-compose.yml
@@ -1,89 +1,0 @@
-version: '3.5'
-
-services:
-  cluster-server:
-    build:
-      context: .
-      dockerfile: Dockerfile
-    restart: "no"
-    container_name: cluster-server
-    environment:
-      CENTRAL_SERVER_LINK: http://multicluster-central-api:5000
-      PORT: 5100
-      HOST_IP: cluster-server
-      TEST_PORT: 5100
-      CLUSTER_NAME: cluster-server
-      HOST: '0.0.0.0'
-    ports:
-      - 5141:5100
-      - 5203:5201
-
-    command: ["python", "server.py", "-v"]
-
-    external_links:
-      - multicluster-central-api
-    networks:
-      - multicluster
-    volumes:
-      - muticlusterserver:/data
-
-  cluster-server-2:
-    build:
-      context: .
-      dockerfile: Dockerfile
-    restart: "no"
-    container_name: cluster-server-2
-    environment:
-      CENTRAL_SERVER_LINK: http://multicluster-central-api:5000
-      PORT: 5100
-      HOST_IP: cluster-server-2
-      TEST_PORT: 5101
-      CLUSTER_NAME: cluster-server-2
-      HOST: '0.0.0.0'
-    ports:
-      - 5142:5100
-      - 5202:5201
-
-    command: ["python", "server.py", "-v"] 
-
-    external_links:
-      - multicluster-central-api
-    networks:
-      - multicluster
-    volumes:
-      - muticlusterserver2:/data
-
-  cluster-server-3:
-    build:
-      context: .
-      dockerfile: Dockerfile
-    restart: "no"
-    container_name: cluster-server-3
-    environment:
-      CENTRAL_SERVER_LINK: http://multicluster-central-api:5000
-      PORT: 5100
-      HOST_IP: cluster-server-3
-      TEST_PORT: 5102
-      CLUSTER_NAME: cluster-server-3
-      HOST: '0.0.0.0'
-    ports:
-      - 5144:5100
-      - 5201:5201
-
-    command: ["python", "server.py", "-v"]
-
-    external_links:
-      - multicluster-central-api
-    networks:
-      - multicluster
-    volumes:
-      - muticlusterserver3:/data
-  
-networks:
-  multicluster:
-    external: true
-
-volumes:
-  muticlusterserver:
-  muticlusterserver2:
-  muticlusterserver3:
