+++ viewServer/scripts/start.sh
